Move impl. of ank get workloads into own file
Issue-Id: #229
